Phase 23: phpstan 8 clean Jobs, Listeners, Providers and expand coverage
- Type-safe fixes in app/Jobs (FireEvent, GenerateCoverThumbnail, GenerateTemporaryInvite, LoadTorrentBoughtUsers, ManagePlugin, RemoveUserVipStatus, SaveIpLogCacheToDB, SeedBonusJob, SendLoginNotify).
- Type-safe fixes in app/Listeners (ClearTorrentCache, DeductUserBonusWhenTorrentDeleted, RemoveOauthTokens, SendEmailNotificationWhenTorrentCreated, SyncTorrentToElasticsearch, SyncTorrentToMeilisearch, TestTorrentUpdated).
- Provider fixes: replace env() with Env::get() in AppServiceProvider, request->cookies->all() in AuthServiceProvider, EventServiceProvider  type, AppPanelProvider boot return type.
- Add app/Jobs, app/Listeners, app/Providers, app/ValueObjects, app/DTOs to phpstan level6/7/8 paths; baselines remain empty.

+++ app/Jobs/RemoveUserVipStatus.php
@@ -41,8 +41,8 @@ class RemoveUserVipStatus
             $message = [];
             $user->vip_added = 'no';
             $user->vip_until = null;
-            if ($user->class <= User::CLASS_VIP) {
-                $user->class = User::CLASS_USER;
+            if ($user->class <= (int) User::CLASS_VIP) {
+                $user->class = (int) User::CLASS_USER;
                 $subject = \App\Support\Locale::trans("cleanup.msg_vip_status_removed", [], $locale);
                 $msg = \App\Support\Locale::trans("cleanup.msg_vip_status_removed_body", [], $locale);
                 $message = [
